requirements.txt
fake_useragent==0.1.11
lxml==4.9.1
requests==2.28.1
PyYAML==6.0
